# Optimized erase_c with a big index