Little extra cost for a small list or a small index.

---

```cpp
using list = brigand::erase_c<brigand::range<int, 0, SIZE>, 0>;
```

Before:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.05s - 35M | 00.06s - 36M | 00.06s - 36M | 00.07s - 36M | 00.07s - 37M
gcc | 00.04s - 19M | 00.04s - 19M | 00.05s - 19M | 00.05s - 20M | 00.06s - 21M

After:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.06s - 36M | 00.07s - 36M | 00.07s - 37M | 00.07s - 37M | 00.08s - 38M
gcc | 00.05s - 20M | 00.05s - 20M | 00.05s - 20M | 00.06s - 21M | 00.07s - 22M

---

```cpp
using list = brigand::erase_c<brigand::range<int, 0, SIZE>, SIZE/2>;
```

Before:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.05s - 35M | 00.06s - 36M | 00.07s - 38M | 00.10s - 42M | 00.19s - 58M
gcc | 00.04s - 19M | 00.05s - 20M | 00.06s - 21M | 00.08s - 26M | 00.17s - 43M

After:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.06s - 36M | 00.07s - 36M | 00.07s - 37M | 00.08s - 38M | 00.10s - 40M
gcc | 00.04s - 20M | 00.05s - 20M | 00.05s - 20M | 00.06s - 21M | 00.08s - 24M

---

```cpp
using list = brigand::erase_c<brigand::range<int, 0, SIZE>, SIZE-1>;
```

Before:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.06s - 35M | 00.07s - 37M | 00.08s - 39M | 00.13s - 48M | ko
gcc | 00.04s - 19M | 00.05s - 20M | 00.06s - 23M | 00.11s - 32M | 00.28s - 65M

After:

Size= | 10 | 50 | 100 | 200 | 400
--- | --- | --- | --- | --- | ---
clang | 00.06s - 36M | 00.07s - 37M | 00.07s - 37M | 00.08s - 38M | 00.10s - 42M
gcc | 00.05s - 20M | 00.05s - 20M | 00.06s - 20M | 00.06s - 22M | 00.09s - 26M